NAS-121735: Move some things around and remove comments
